Add feature detection code and a few digit images
Feature detection is being done right now using Brute Force Matching
(BFMatching). The identified digit is being selected as the match with the
minimum mean value for the "distance" value returned by the feature
matching. This needs improvements and does not work every time. Some other
variable needs to be incorporated in the future.