Add helper functions for cropping and optimizing contours
Includes function for cropping an image based on the extremas of a given
contour. New function for selecting the best contour to use for character
recognition. This function needs to be modified for efficiency in the
future. Currently, it makes repeated calls to findContours every time the
image is manipulated.