Add file for comparing contours using Hausdorff Distance Extractor
This is a breakthrough for the project's goal of character recognition.
The code finds the optimal contours from the two images being compared,
and then passes them through OpenCV's
cv2.createHausdorffDistanceExtractor(). The smaller the value returned,
the more similar the two images are. This was tested with above average
reliability for the image files for fours and fives.